# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "hoshiyomi API",
    "description": "hoshiyomi 后端 REST API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/auth/jwt/echo": {
      "get": {
        "tags": [
          "auth"
        ],
        "operationId": "Auth__jwtEcho",
        "responses": {
          "200": {
            "description": "认证检查示例",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_JwtEchoResp"
                }
              }
            }
          },
          "401": {
            "description": "令牌无效"
          }
        }
      }
    },
    "/api/v1/auth/jwt/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "Auth__jwtLogin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_LoginResp"
                }
              }
            }
          },
          "400": {
            "description": "参数或校验错误"
          },
          "403": {
            "description": "用户名或密码错误"
          }
        }
      }
    },
    "/api/v1/auth/jwt/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "Auth__jwtLogout",
        "responses": {
          "200": {
            "description": "已撤销全部刷新令牌",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_MessageResp"
                }
              }
            }
          },
          "401": {
            "description": "令牌无效"
          }
        }
      }
    },
    "/api/v1/auth/jwt/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "operationId": "Auth__jwtMe",
        "responses": {
          "200": {
            "description": "当前用户",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_AuthStateResp"
                }
              }
            }
          },
          "401": {
            "description": "令牌无效"
          }
        }
      }
    },
    "/api/v1/auth/jwt/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "Auth__jwtRefresh",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "新令牌对",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_RefreshResp"
                }
              }
            }
          },
          "400": {
            "description": "参数或校验错误"
          },
          "401": {
            "description": "刷新令牌无效或已过期"
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "Auth__login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "登录成功,设置会话 Cookie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_AuthStateResp"
                }
              }
            }
          },
          "400": {
            "description": "参数或校验错误"
          },
          "403": {
            "description": "用户名或密码错误"
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "Auth__logout",
        "responses": {
          "200": {
            "description": "已退出",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_MessageResp"
                }
              }
            }
          },
          "401": {
            "description": "未登录"
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "operationId": "Auth__me",
        "responses": {
          "200": {
            "description": "当前用户",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_AuthStateResp"
                }
              }
            }
          },
          "401": {
            "description": "未登录"
          }
        }
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "Auth__register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "注册的用户",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_UserResp"
                }
              }
            }
          },
          "400": {
            "description": "参数或校验错误"
          },
          "409": {
            "description": "用户名或邮箱已存在"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiResponse_AuthStateResp": {
        "type": "object",
        "description": "统一响应信封:code=0 成功,非 0 为 HTTP 状态码;data 恒序列化(错误时为 null)",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "description": "0 表示成功,否则为失败的 HTTP 状态码"
          },
          "data": {
            "type": "object",
            "required": [
              "user",
              "permissions"
            ],
            "properties": {
              "permissions": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Perm"
                }
              },
              "user": {
                "$ref": "#/components/schemas/UserResp"
              }
            }
          },
          "errors": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ValidationErrors",
                "description": "校验失败明细,仅在校验失败时出现"
              }
            ]
          },
          "message": {
            "type": "string",
            "description": "简短的可读消息;成功时为 \"ok\""
          }
        }
      },
      "ApiResponse_JwtEchoResp": {
        "type": "object",
        "description": "统一响应信封:code=0 成功,非 0 为 HTTP 状态码;data 恒序列化(错误时为 null)",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "description": "0 表示成功,否则为失败的 HTTP 状态码"
          },
          "data": {
            "type": "object",
            "required": [
              "user_id",
              "username"
            ],
            "properties": {
              "user_id": {
                "type": "integer",
                "format": "int64",
                "minimum": 0
              },
              "username": {
                "type": "string"
              }
            }
          },
          "errors": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ValidationErrors",
                "description": "校验失败明细,仅在校验失败时出现"
              }
            ]
          },
          "message": {
            "type": "string",
            "description": "简短的可读消息;成功时为 \"ok\""
          }
        }
      },
      "ApiResponse_LoginResp": {
        "type": "object",
        "description": "统一响应信封:code=0 成功,非 0 为 HTTP 状态码;data 恒序列化(错误时为 null)",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "description": "0 表示成功,否则为失败的 HTTP 状态码"
          },
          "data": {
            "type": "object",
            "required": [
              "access_token",
              "refresh_token",
              "state"
            ],
            "properties": {
              "access_token": {
                "type": "string"
              },
              "refresh_token": {
                "type": "string"
              },
              "state": {
                "$ref": "#/components/schemas/AuthStateResp"
              }
            }
          },
          "errors": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ValidationErrors",
                "description": "校验失败明细,仅在校验失败时出现"
              }
            ]
          },
          "message": {
            "type": "string",
            "description": "简短的可读消息;成功时为 \"ok\""
          }
        }
      },
      "ApiResponse_MessageResp": {
        "type": "object",
        "description": "统一响应信封:code=0 成功,非 0 为 HTTP 状态码;data 恒序列化(错误时为 null)",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "description": "0 表示成功,否则为失败的 HTTP 状态码"
          },
          "data": {
            "type": "object",
            "required": [
              "message"
            ],
            "properties": {
              "message": {
                "type": "string"
              }
            }
          },
          "errors": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ValidationErrors",
                "description": "校验失败明细,仅在校验失败时出现"
              }
            ]
          },
          "message": {
            "type": "string",
            "description": "简短的可读消息;成功时为 \"ok\""
          }
        }
      },
...
      "ApiResponse_RefreshResp": {
        "type": "object",
        "description": "统一响应信封:code=0 成功,非 0 为 HTTP 状态码;data 恒序列化(错误时为 null)",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "description": "0 表示成功,否则为失败的 HTTP 状态码"
          },
          "data": {
            "type": "object",
            "required": [
              "access_token",
              "refresh_token"
            ],
            "properties": {
              "access_token": {
                "type": "string"
              },
              "refresh_token": {
                "type": "string"
              }
            }
          },
          "errors": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ValidationErrors",
                "description": "校验失败明细,仅在校验失败时出现"
              }
            ]
          },
          "message": {
            "type": "string",
            "description": "简短的可读消息;成功时为 \"ok\""
          }
        }
      },
...
      "ApiResponse_UserResp": {
        "type": "object",
        "description": "统一响应信封:code=0 成功,非 0 为 HTTP 状态码;data 恒序列化(错误时为 null)",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "description": "0 表示成功,否则为失败的 HTTP 状态码"
          },
          "data": {
            "type": "object",
            "required": [
              "id",
              "username",
              "email",
              "created_at",
              "updated_at"
            ],
            "properties": {
              "created_at": {
                "type": "string"
              },
              "email": {
                "type": "string"
              },
              "id": {
                "type": "integer",
                "format": "int64",
                "minimum": 0
              },
              "updated_at": {
                "type": "string"
              },
              "username": {
                "type": "string"
              }
            }
          },
          "errors": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ValidationErrors",
                "description": "校验失败明细,仅在校验失败时出现"
              }
            ]
          },
          "message": {
            "type": "string",
            "description": "简短的可读消息;成功时为 \"ok\""
          }
        }
      },
      "AuthStateResp": {
        "type": "object",
        "required": [
          "user",
          "permissions"
        ],
        "properties": {
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Perm"
            }
          },
          "user": {
            "$ref": "#/components/schemas/UserResp"
          }
        }
      },
...
      "FieldViolation": {
        "type": "object",
        "description": "某个字段的一条失败规则",
        "required": [
          "code",
          "params"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "失败的规则名(如 \"length\"、\"email\")"
          },
          "message": {
            "type": [
              "string",
              "null"
            ],
            "description": "DTO 为该规则声明的消息(声明了才有)"
          },
          "params": {
            "type": "object",
            "description": "失败规则的参数(如 {\"min\": 3, \"max\": 20})。提交的字段值刻意不回显:validator 会把它记为 params.value,原样回显会把原始输入(口令、令牌)返回给客户端。",
            "additionalProperties": true
          }
        }
      },
...
      "LoginReq": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
...
      "Perm": {
        "type": "string",
        "description": "权限点:serde 与 strum 序列化码必须一致(见 tests::test_permissions_consistency)",
        "enum": [
          "*",
          "user:read",
          "user:write",
          "user:delete",
          "user:*",
          "role:read",
          "role:write",
          "role:delete",
          "role:*"
        ]
      },
      "RefreshReq": {
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "properties": {
          "refresh_token": {
            "type": "string"
          }
        }
      },
...
      "RegisterReq": {
        "type": "object",
        "required": [
          "username",
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
...
      "UserResp": {
        "type": "object",
        "required": [
          "id",
          "username",
          "email",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "updated_at": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "ValidationErrors": {
        "type": "object",
        "description": "按字段分组的全部违规。键为扁平化的字段路径:顶层字段是 email,嵌套结构体字段是 profile.email,列表元素是 items[0].name。映射有序,故序列化结果稳定。",
        "additionalProperties": {
          "type": "array",
          "items": {
            "$ref": "#/components/schemas/FieldViolation"
          }
        },
        "propertyNames": {
          "type": "string"
        }
      }
    }
  }
}